LayoutTests/fast/profiler/many-calls-in-the-same-scope.html

<html>
<head>
<script src="resources/profiler-test-JS-resources.js"></script>
<script>
if (window.testRunner)
    testRunner.dumpAsText();
if (window.internals)
    internals.settings.setJavaScriptProfilingEnabled(true);

function startTest()
{
    console.profile("Many Calls In The Same Scope");

    insertNewText();
    insertGivenText("This was a triumph.");
    arrayOperatorFunction(7);
    intermediaryFunction();
    anonymousFunction();
    end()
    endT();
    endT();

    endTest();
}

function end()
{
    var x = 0;
}

function endT()
{
    var y = 1;
}

</script>
</head>

<body onload="startTest()">
This page's JavaScript has many function calls in the same scope.
<br>
<br>
To run this test manually, load it in the browser then load the WebInspector and look at
the profile.  In the profile many functions should be the children of startTest.
Use the sorting capabilites to make sure the similarly named functions are sorted
correctly.
<div id="output"></div>
</body>
</html>
